requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0

requests==2.31.0
structlog==22.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.22

google-cloud-storage==2.13.0